ExampleSpear merge & improvements
