+++ ExampleMod/Items/Weapons/ExampleSpear.cs
@@ -1,4 +1,6 @@
-﻿using ExampleMod.Projectiles;
+﻿using System;
+using ExampleMod.Projectiles;
+using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
@@ -24,18 +26,116 @@ namespace ExampleMod.Items.Weapons
 			item.height = 32;
 			item.scale = 1f;
 			item.rare = 5;
-			item.UseSound = SoundID.Item1;
-			item.shoot = mod.ProjectileType<ExampleSpearProjectile>();
-			item.value = 1000;
+			item.value = Item.sellPrice(silver: 10);
+
+			item.melee = true;
 			item.noMelee = true; // Important because the spear is actually a projectile instead of an item. This prevents the melee hitbox of this item.
 			item.noUseGraphic = true; // Important, it's kind of wired if people see two spears at one time. This prevents the melee animation of this item.
-			item.melee = true;
-			item.autoReuse = true; // Most spears don't autoReuse, but it's possible
+			item.autoReuse = true; // Most spears don't autoReuse, but it's possible when used in conjunction with CanUseItem()
+
+			item.UseSound = SoundID.Item1;
+			item.shoot = mod.ProjectileType<ExampleSpearProjectile>();
 		}
 
 		public override bool CanUseItem(Player player)
 		{
-			return player.ownedProjectileCounts[item.shoot] < 1; // This is to ensure the spear doesn't bug out when using autoReuse = true
+			// Ensures no more than one spear can be thrown out, use this when using autoReuse
+			return player.ownedProjectileCounts[item.shoot] < 1; 
+		}
+	}
+
+	public class ExampleSpearProjectile : ModProjectile
+	{
+		public override void SetStaticDefaults()
+		{
+			DisplayName.SetDefault("Spear");
+		}
+
+		public override void SetDefaults()
+		{
+			projectile.width = 18;
+			projectile.height = 18;
+			projectile.aiStyle = 19;
+			projectile.penetrate = -1;
+			projectile.scale = 1.3f;
+			projectile.alpha = 0;
+
+			projectile.hide = true;
+			projectile.ownerHitCheck = true;
+			projectile.melee = true;
+			projectile.tileCollide = false;
+			projectile.friendly = true;
+		}
+
+		// In here the AI uses this example, to make the code more organized and readable
+		// Also showcased in ExampleJavelinProjectile.cs
+		public float movementFactor // Change this value to alter how fast the spear moves
+		{
+			get { return projectile.ai[0]; }
+			set { projectile.ai[0] = value; }
+		}
+
+		// It appears that for this AI, only the ai0 field is used!
+		public override void AI()
+		{
+			// Since we access the owner player instance so much, it's useful to create a helper local variable for this
+			// Sadly, Projectile/ModProjectile does not have its own
+			Player projOwner = Main.player[projectile.owner];
+			// Here we set some of the projectile's owner properties, such as held item and itemtime, along with projectile direction and position based on the player
+			Vector2 ownerMountedCenter = projOwner.RotatedRelativePoint(projOwner.MountedCenter, true);
+			projectile.direction = projOwner.direction;
+			projOwner.heldProj = projectile.whoAmI;
+			projOwner.itemTime = projOwner.itemAnimation;
+			projectile.position.X = ownerMountedCenter.X - (float)(projectile.width / 2);
+			projectile.position.Y = ownerMountedCenter.Y - (float)(projectile.height / 2);
+			// As long as the player isn't frozen, the spear can move
+			if (!projOwner.frozen)
+			{
+				if (movementFactor == 0f) // When initially thrown out, the ai0 will be 0f
+				{
+					movementFactor = 3f; // Make sure the spear moves forward when initially thrown out
+					projectile.netUpdate = true; // Make sure to netUpdate this spear
+				}
+				if (projOwner.itemAnimation < projOwner.itemAnimationMax / 3) // Somewhere along the item animation, make sure the spear moves back
+				{
+					movementFactor -= 2.4f;
+				}
+				else // Otherwise, increase the movement factor
+				{
+					movementFactor += 2.1f;
+				}
+			}
+			// Change the spear position based off of the velocity and the movementFactor
+			projectile.position += projectile.velocity * movementFactor;
+			// When we reach the end of the animation, we can kill the spear projectile
+			if (projOwner.itemAnimation == 0)
+			{
+				projectile.Kill();
+			}
+			// Apply proper rotation, with an offset of 135 degrees due to the sprite's rotation, notice the usage of MathHelper, use this class!
+			// MathHelper.ToRadians(xx degrees here)
+			projectile.rotation = projectile.velocity.ToRotation() + MathHelper.ToRadians(135f);
+			// Offset by 90 degrees here
+			if (projectile.spriteDirection == -1)
+			{
+				projectile.rotation -= MathHelper.ToRadians(90f);
+			}
+
+			// These dusts are added later, for the 'ExampleMod' effect
+			if (Main.rand.Next(3) == 0)
+			{
+				Dust dust = Dust.NewDustDirect(projectile.position, projectile.height, projectile.width, mod.DustType<Dusts.Sparkle>(),
+					projectile.velocity.X * .2f, projectile.velocity.Y * .2f, 200, Scale: 1.2f);
+				dust.velocity += projectile.velocity * 0.3f;
+				dust.velocity *= 0.2f;
+			}
+			if (Main.rand.Next(4) == 0)
+			{
+				Dust dust = Dust.NewDustDirect(projectile.position, projectile.height, projectile.width, mod.DustType<Dusts.Sparkle>(),
+					0, 0, 254, Scale: 0.3f);
+				dust.velocity += projectile.velocity * 0.5f;
+				dust.velocity *= 0.5f;
+			}
 		}
 	}
 }
