--- ExampleMod/Projectiles/ExampleSpearProjectile.cs
@@ -1,104 +1,0 @@
-﻿using Microsoft.Xna.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Terraria;
-using Terraria.ID;
-using Terraria.ModLoader;
-
-namespace ExampleMod.Projectiles
-{
-	public class ExampleSpearProjectile : ModProjectile
-	{
-		public override void SetStaticDefaults()
-		{
-			DisplayName.SetDefault("Spear");
-		}
-
-		public override void SetDefaults()
-		{
-			projectile.width = 18;
-			projectile.height = 18;
-			projectile.aiStyle = 19;
-			projectile.friendly = true;
-			projectile.penetrate = -1;
-			projectile.tileCollide = false;
-			projectile.scale = 1.3f;
-			projectile.hide = true;
-			projectile.ownerHitCheck = true;
-			projectile.melee = true;
-			projectile.alpha = 0;
-		}
-
-		// In here the AI uses this example, to make the code more organized and readable
-		// Also showcased in ExampleJavelinProjectile.cs
-		public float movementFactor // Change this value to alter how fast the spear moves
-		{
-			get { return projectile.ai[0]; }
-			set { projectile.ai[0] = value; }
-		}
-		// It appears that for this AI, only the ai0 field is used!
-
-		public override void AI()
-		{
-			// Since we access the owner player instance so much, it's useful to create a helper local variable for this
-			// Sadly, Projectile/ModProjectile does not have its own
-			Player projOwner = Main.player[projectile.owner];
-			// Here we set some of the projectile's owner properties, such as held item and itemtime, along with projectile direction and position based on the player
-			Vector2 ownerMountedCenter = projOwner.RotatedRelativePoint(projOwner.MountedCenter, true);
-			projectile.direction = projOwner.direction;
-			projOwner.heldProj = projectile.whoAmI;
-			projOwner.itemTime = projOwner.itemAnimation;
-			projectile.position.X = ownerMountedCenter.X - (float)(projectile.width / 2);
-			projectile.position.Y = ownerMountedCenter.Y - (float)(projectile.height / 2);
-			// As long as the player isn't frozen, the spear can move
-			if (!projOwner.frozen)
-			{
-				if (movementFactor == 0f) // When initially thrown out, the ai0 will be 0f
-				{
-					movementFactor = 3f; // Make sure the spear moves forward when initially thrown out
-					projectile.netUpdate = true; // Make sure to netUpdate this spear
-				}
-				if (projOwner.itemAnimation < projOwner.itemAnimationMax / 3) // Somewhere along the item animation, make sure the spear moves back
-				{
-					movementFactor -= 2.4f;
-				}
-				else // Otherwise, increase the movement factor
-				{
-					movementFactor += 2.1f;
-				}
-			}
-			// Change the spear position based off of the velocity and the movementFactor
-			projectile.position += projectile.velocity * movementFactor;
-			// When we reach the end of the animation, we can kill the spear projectile
-			if (projOwner.itemAnimation == 0)
-			{
-				projectile.Kill();
-			}
-			// Apply proper rotation, with an offset of 135 degrees due to the sprite's rotation, notice the usage of MathHelper, use this class!
-			// MathHelper.ToRadians(xx degrees here)
-			projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + MathHelper.ToRadians(135f);
-			// Offset by 90 degrees here
-			if (projectile.spriteDirection == -1)
-			{
-				projectile.rotation -= MathHelper.ToRadians(90f);
-			}
-
-			// These dusts are added later, for the 'ExampleMod' effect
-			if (Main.rand.Next(3) == 0)
-			{
-				int dustIndex = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, mod.DustType<Dusts.Sparkle>(), projectile.velocity.X * 0.2f, projectile.velocity.Y * 0.2f, 200, default(Color), 1.2f);
-				Main.dust[dustIndex].velocity += projectile.velocity * 0.3f;
-				Main.dust[dustIndex].velocity *= 0.2f;
-			}
-			if (Main.rand.Next(4) == 0)
-			{
-				int dustIndex = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, mod.DustType<Dusts.Smoke>(), 0f, 0f, 254, default(Color), 0.3f);
-				Main.dust[dustIndex].velocity += projectile.velocity * 0.5f;
-				Main.dust[dustIndex].velocity *= 0.5f;
-				return;
-			}
-		}
-	}
-}
